seprate schema to 2 files. Product.js and Setup.js
