+++ server/mongodb/models/setup.js
@@ -9,16 +9,6 @@ const setupSchema = new mongoose.Schema(
   { timestamps: true }
 );
 
-const productSchema = new mongoose.Schema(
-  {
-    type: { type: String, required: true },
-    brand: { type: String, required: true },
-    model: { type: String, required: true },
-  },
-  { timestamps: true }
-);
-
 const Setup = mongoose.model('Setup', setupSchema);
-const Product = mongoose.model('Product', productSchema);
 
-export default { Setup, Product };
+export default { Setup };
